Deleted usePlayers hook + Refactored responsive layout
