+++ apps/frontend/src/components/players/PlayerTableActionRow.tsx
@@ -1,8 +1,8 @@
 import {
     Box,
     Button,
+    Card,
     Flex,
-    IconButton,
     Separator,
     Text,
     TextField,
@@ -11,75 +11,149 @@ import AddPlayerDialog from "./AddPlayerDialog.tsx";
 import { DownloadIcon, SearchIcon, TrashIcon, UploadIcon } from "lucide-react";
 import type { Player } from "../../utils/types.ts";
 import type { Table } from "@tanstack/react-table";
-import type { PlayerSchema } from "../../utils/schemas.ts";
-import { z } from "zod";
 import { type ChangeEvent, useRef } from "react";
+import { doc, writeBatch } from "firebase/firestore";
+import { clientFirestore } from "../../utils/firebase.ts";
+import { useUser } from "../../hooks/useUser.ts";
 
 type PlayerTableActionRowProps = {
     table: Table<Player>;
-    selectedPlayerIds: string[];
-    addPlayer: (player?: z.infer<typeof PlayerSchema>) => Promise<void>;
-    exportJSON: () => void;
-    importJSON: (file: Blob) => void;
-    handleRemovePlayers: () => void;
+    players: Player[];
 };
 
 export default function PlayerTableActionRow({
     table,
-    selectedPlayerIds,
-    addPlayer,
-    exportJSON,
-    importJSON,
-    handleRemovePlayers,
+    players,
 }: PlayerTableActionRowProps) {
+    const { user } = useUser();
+
     const fileInputRef = useRef<HTMLInputElement | null>(null);
 
-    const onChange = (event: ChangeEvent<HTMLInputElement>) => {
+    const exportJSON = () => {
+        const jsonData = new Blob([JSON.stringify(players)], {
+            type: "application/json",
+        });
+
+        const jsonURL = URL.createObjectURL(jsonData);
+
+        const link = document.createElement("a");
+
+        link.href = jsonURL;
+        link.download = "players.json";
+        document.body.appendChild(link);
+        link.click();
+        document.body.removeChild(link);
+    };
+
+    const selectedPlayerIds = table
+        .getFilteredSelectedRowModel()
+        .rows.map((row) => (row.original as Player).id);
+
+    const handleRemovePlayers = async () => {
+        if (!user) {
+            throw new Error("User not authenticated");
+        }
+
+        const batch = writeBatch(clientFirestore);
+        selectedPlayerIds.forEach((selectedPlayerId) => {
+            batch.delete(
+                doc(
+                    clientFirestore,
+                    `users/${user.uid!}/players/${selectedPlayerId}`,
+                ),
+            );
+        });
+
+        await batch.commit();
+        table.resetRowSelection();
+    };
+
+    const importJSON = (event: ChangeEvent<HTMLInputElement>) => {
+        if (!user) {
+            throw new Error("User not authenticated");
+        }
+
         if (event.target.files) {
-            importJSON(event.target.files[0]);
+            const fileReader = new FileReader();
+
+            fileReader.onload = async (event) => {
+                if (event.target) {
+                    const deletePlayerBatch = writeBatch(clientFirestore);
+
+                    players.forEach((player) => {
+                        deletePlayerBatch.delete(
+                            doc(
+                                clientFirestore,
+                                `users/${user.uid!}/players/${player.id!}`,
+                            ),
+                        );
+                    });
+
+                    const addPlayerBatch = writeBatch(clientFirestore);
+
+                    const parsedPlayers: Player[] = JSON.parse(
+                        event.target.result as string,
+                    );
+
+                    parsedPlayers.forEach((player) => {
+                        addPlayerBatch.set(
+                            doc(
+                                clientFirestore,
+                                `users/${user.uid!}/players`,
+                                player.id,
+                            ),
+                            {
+                                name: player.name,
+                                number: player.number,
+                                position: player.position,
+                                availabilities: player.availabilities,
+                            },
+                        );
+                    });
+
+                    await deletePlayerBatch.commit();
+
+                    await addPlayerBatch.commit();
+                }
+            };
+
+            fileReader.readAsText(event.target.files[0]);
         }
     };
 
     return (
-        <Flex justify="between" align="center" mb="5">
-            <Flex align="center" gap="4">
-                <AddPlayerDialog addPlayer={addPlayer} />
-                <Box
-                    display={{
-                        initial: "none",
-                        sm: "block",
-                    }}
-                >
+        <>
+            <Flex
+                direction={{
+                    initial: "column",
+                    sm: "row",
+                }}
+                justify="between"
+                align={{
+                    initial: "start",
+                    sm: "center",
+                }}
+                gap={{
+                    initial: "4",
+                    sm: "0",
+                }}
+                mb="4"
+            >
+                <Flex align="center" gap="2" wrap="wrap">
+                    <AddPlayerDialog user={user} />
                     <Button variant="outline" onClick={exportJSON}>
                         <DownloadIcon size={15} />
                         Export JSON
                     </Button>
-                </Box>
-                <Box
-                    display={{
-                        initial: "block",
-                        sm: "none",
-                    }}
-                >
-                    <IconButton variant="outline" onClick={exportJSON}>
-                        <DownloadIcon size={15} />
-                    </IconButton>
-                </Box>
-                <input
-                    id="import-json"
-                    ref={fileInputRef}
-                    name="import-json"
-                    type="file"
-                    accept=".json, application/json"
-                    onChange={onChange}
-                    style={{ display: "none" }}
-                />
-                <Box
-                    display={{
-                        initial: "none",
-                        sm: "block",
-                    }}
-                >
+                    <input
+                        id="import-json"
+                        ref={fileInputRef}
+                        name="import-json"
+                        type="file"
+                        accept=".json, application/json"
+                        onChange={importJSON}
+                        style={{ display: "none" }}
+                    />
                     <label htmlFor="import-json">
                         <Button
                             variant="outline"
@@ -89,85 +163,48 @@ export default function PlayerTableActionRow({
                             Import JSON
                         </Button>
                     </label>
-                    {selectedPlayerIds.length > 0 && (
-                        <>
-                            <Separator size="2" orientation="vertical" />
-                            <Text>{selectedPlayerIds.length} selected</Text>
-                            <Button color="red" onClick={handleRemovePlayers}>
-                                <TrashIcon size={15} />
-                                Delete Selected
-                            </Button>
-                        </>
-                    )}
-                </Box>
+                </Flex>
                 <Box
-                    display={{
-                        initial: "block",
-                        sm: "none",
+                    width={{
+                        initial: "100%",
+                        sm: "250px",
                     }}
                 >
-                    <label htmlFor="import-json">
-                        <IconButton
-                            variant="outline"
-                            onClick={() => fileInputRef.current?.click()}
-                        >
-                            <UploadIcon size={15} />
-                        </IconButton>
-                    </label>
+                    <TextField.Root
+                        id="search"
+                        name="search"
+                        placeholder="Search players by name"
+                        value={
+                            (table
+                                .getColumn("name")
+                                ?.getFilterValue() as string) ?? ""
+                        }
+                        onChange={(event) =>
+                            table
+                                .getColumn("name")
+                                ?.setFilterValue(event.target.value)
+                        }
+                    >
+                        <TextField.Slot>
+                            <SearchIcon size={15} />
+                        </TextField.Slot>
+                    </TextField.Root>
                 </Box>
             </Flex>
-            <Box
-                width="200px"
-                display={{
-                    initial: "none",
-                    xs: "block",
-                }}
-            >
-                <TextField.Root
-                    id="search"
-                    name="search"
-                    placeholder="Search players by name"
-                    value={
-                        (table.getColumn("name")?.getFilterValue() as string) ??
-                        ""
-                    }
-                    onChange={(event) =>
-                        table
-                            .getColumn("name")
-                            ?.setFilterValue(event.target.value)
-                    }
-                >
-                    <TextField.Slot>
-                        <SearchIcon size={15} />
-                    </TextField.Slot>
-                </TextField.Root>
-            </Box>
-            <Box
-                width="90px"
-                display={{
-                    initial: "block",
-                    xs: "none",
-                }}
-            >
-                <TextField.Root
-                    id="search"
-                    name="search"
-                    placeholder="Search"
-                    value={
-                        (table.getColumn("name")?.getFilterValue() as string) ??
-                        ""
-                    }
-                    onChange={(event) =>
-                        table
-                            .getColumn("name")
-                            ?.setFilterValue(event.target.value)
-                    }
-                >
-                    <TextField.Slot>
-                        <SearchIcon size={15} />
-                    </TextField.Slot>
-                </TextField.Root>
-            </Box>
-        </Flex>
+            {selectedPlayerIds.length > 0 && (
+                <Card mb="4">
+                    <Flex justify="start" align="center" gap="2" wrap="wrap">
+                        <Text color="gray" weight="medium" size="2" wrap="nowrap">
+                            {selectedPlayerIds.length} players selected
+                        </Text>
+                        <Separator orientation="vertical" size="1" />
+                        <Button color="red" onClick={handleRemovePlayers}>
+                            <TrashIcon size={15} />
+                            Delete Selected
+                        </Button>
+                    </Flex>
+                </Card>
+            )}
+        </>
     );
 }
