+++ apps/frontend/src/pages/create-an-account/CreateAnAccount.tsx
@@ -120,7 +120,7 @@ export default function CreateAnAccountCard() {
                             </Text>
                         )}
                     </Box>
-                    <Flex direction="row-reverse" gap="4">
+                    <Flex direction="row-reverse" gap="2">
                         <Button
                             type="submit"
                             disabled={isSubmitting || isValidating}
