+++ apps/frontend/src/pages/signin/SignIn.tsx
@@ -103,7 +103,7 @@ export default function SignIn() {
                             </Text>
                         )}
                     </Box>
-                    <Flex direction="row-reverse" gap="4">
+                    <Flex direction="row-reverse" gap="2">
                         <Button
                             type="submit"
                             disabled={isSubmitting || isValidating}
